Refactor : 생성자 인자를 List<Integer> 에서 List<LottoNumber> 로 변경 Co-authored-by: worldii <bbc2788@naver.com>
