+++ src/test/java/lotto/model/LottoTest.java
@@ -1,6 +1,5 @@
-package lotto;
+package lotto.model;
 
-import lotto.model.Lotto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -8,6 +7,7 @@ import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
 import java.util.List;
+import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
@@ -36,14 +36,14 @@ public class LottoTest {
     @MethodSource("createLotto")
     @DisplayName("여러 로또 넘버가 중복되지 않는지 테스트")
     void 로또_중복(List<Integer> numbers) {
-        assertThatCode(() -> new Lotto(numbers))
+        assertThatCode(() -> new Lotto(createLottoNumbers(numbers)))
                 .doesNotThrowAnyException();
     }
 
     @Test
     @DisplayName("여러 로또 넘버가 중복되는 않는지 테스트")
     void 로또_중복() {
-        assertThatCode(() -> new Lotto(List.of(1, 1, 2, 3, 4, 5)))
+        assertThatCode(() -> new Lotto(createLottoNumbers(List.of(1, 1, 2, 3, 4, 5))))
                 .isInstanceOf(IllegalStateException.class)
                 .hasMessage("로또 넘버는 중복되면 안됩니다.");
     }
@@ -52,7 +52,7 @@ public class LottoTest {
     @MethodSource("createLottoLengthFail")
     @DisplayName("로또 번호는 6개가 아니면 예외를 던진다")
     void 로또_번호_6개가_아니면_예외_던진다(List<Integer> numbers) {
-        assertThatCode(() -> new Lotto(numbers))
+        assertThatCode(() -> new Lotto(createLottoNumbers(numbers)))
                 .isInstanceOf(IllegalStateException.class)
                 .hasMessage("로또 번호는 6개이여야 합니다");
     }
@@ -60,9 +60,12 @@ public class LottoTest {
     @Test
     @DisplayName("로또 번호는 6개가 아니면 정상 동작")
     void 로또_번호_6개이면_정상() {
-        assertThatCode(() -> new Lotto(List.of(1, 2, 3, 4, 5, 6)))
+        assertThatCode(() -> new Lotto(createLottoNumbers(List.of(1, 2, 3, 4, 5, 6))))
                 .doesNotThrowAnyException();
     }
 
 
+    private List<LottoNumber> createLottoNumbers(List<Integer> container) {
+        return container.stream().map(LottoNumber::new).collect(Collectors.toList());
+    }
 }
